Added example of layers to docs
