+++ htf/utils.py
@@ -250,103 +250,6 @@ def iter_from_trajectory(nneighbor_cutoff, universe, selection='all', r_cut=10.,
                 axis=1), hoomd_box, 1.0], ts
 
 
-class EDSLayer(tf.keras.layers.Layer):
-    R''' This layer computes and returns the Lagrange multiplier/EDS coupling constant (alpha)
-    to be used as the EDS bias in the simulation. You call the layer on the
-    collective variable at each step to get the current value of alpha.
-
-    :param set_point: The set point value of the collective variable.
-        This is a constant value which is pre-determined by the user and unique to each cv.
-    :param period: Time steps over which the coupling constant is updated.
-        Hoomd time units are used. If period=100 alpha will be updated each 100 time steps.
-    :param learninig_rate: Learninig_rate in the EDS method.
-    :param cv_scale: Used to adjust the units of the bias to Hoomd units.
-    :param name: Name to be used for layer
-    :return: Alpha, the EDS coupling constant.
-    '''
-
-    def __init__(self, set_point, period, learning_rate=1e-2,
-                 cv_scale=1.0, name='eds-layer', **kwargs):
-        if not tf.is_tensor(set_point):
-            set_point = tf.convert_to_tensor(set_point)
-        if set_point.dtype not in (tf.float32, tf.float64):
-            raise ValueError(
-                'EDS only works with floats, not dtype' +
-                str(set_point.dtype))
-        super().__init__(name, dtype=set_point.dtype, **kwargs)
-        self.set_point = set_point
-        self.period = tf.cast(period, tf.int32)
-        self.cv_scale = cv_scale
-        self.learning_rate = learning_rate
-        self.optimizer = tf.compat.v1.train.AdamOptimizer(learning_rate)
-
-    def get_config(self):
-        base = super().get_config()
-        c = {
-            'set_point': self.set_point.numpy(),
-            'period': self.period,
-            'cv_scale': self.cv_scale,
-            'learning_rate': self.learning_rate,
-        }
-        c.update(base)
-        return c
-
-    def build(self, input_shape):
-        # set-up variables
-        self.mean = self.add_weight(
-            initializer=tf.zeros_initializer(), dtype=self.dtype,
-            shape=input_shape, name='{}.mean'.format(self.name),
-                                    trainable=False)
-        self.ssd = self.add_weight(
-            initializer=tf.zeros_initializer(), dtype=self.dtype,
-            shape=input_shape, name='{}.ssd'.format(self.name),
-            trainable=False)
-        self.n = self.add_weight(
-            initializer=tf.zeros_initializer(),
-            shape=input_shape, dtype=tf.int32, name='{}.n'.format(
-                self.name), trainable=False)
-        self.alpha = self.add_weight(initializer=tf.zeros_initializer(
-        ), shape=input_shape, name='{}.a'.format(self.name), dtype=self.dtype)
-
-    @tf.function
-    def call(self, cv):
-        reset_mask = tf.cast((self.n != 0), self.dtype)
-
-        # reset statistics if n is 0
-        self.mean.assign(self.mean * reset_mask)
-        self.ssd.assign(self.ssd * reset_mask)
-
-        # update statistics
-        # do we update? - masked
-        update_mask = tf.cast(self.n > self.period // 2, self.dtype)
-        delta = (cv - self.mean) * update_mask
-        self.mean.assign_add(
-            tf.math.divide_no_nan(
-                delta,
-                tf.cast(self.n - self.period // 2, self.dtype)
-            )
-        )
-
-        self.ssd.assign_add(delta * (cv - self.mean))
-
-        # update grad
-        update_mask = tf.cast(
-            tf.equal(self.n, self.period - 1), self.dtype)
-        gradient = update_mask * -  2 * \
-            (self.mean - self.set_point) * self.ssd / \
-            tf.cast(self.period, self.dtype) / 2 / self.cv_scale
-
-        tf.cond(pred=tf.equal(self.n, self.period - 1),
-                true_fn=lambda: self.optimizer.apply_gradients([(gradient,
-                                                                 self.alpha)]),
-                false_fn=lambda: tf.no_op())
-
-        # update n. Should reset at period
-        self.n.assign((self.n + 1) % self.period)
-
-        return self.alpha
-
-
 def center_of_mass(positions, mapping, box_size, name='center-of-mass'):
     '''Comptue mapping of the given positions ``N x 3` and mapping ``M x N``
     considering PBC. Returns mapped particles.
